+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## 1.0.2 - 2019-08-23
 ### Added
-
+  - Sorting of admin panel table.
 
 ### Changed
   - Resolve minor spelling mistake in tooltip for delete and edit buttons on admin panel.